Trying to get async rendering of mermaid to work
