+++ console/frontend/src/main/frontend/angular.json
@@ -63,7 +63,8 @@
               "angular-datatables/dist/plugins/buttons/angular-datatables.buttons",
               "pace-js",
               "prismjs"
-            ]
+            ],
+            "webWorkerTsConfig": "tsconfig.worker.json"
           },
           "configurations": {
             "production": {
@@ -136,7 +137,8 @@
             "styles": [
               "src/styles.scss"
             ],
-            "scripts": []
+            "scripts": [],
+            "webWorkerTsConfig": "tsconfig.worker.json"
           }
         }
       }
