+++ console/frontend/src/main/frontend/src/app/components/ng-mermaid/ng-mermaid.worker.ts
@@ -1,0 +1,6 @@
+/// <reference lib="webworker" />
+
+addEventListener('message', ({ data }) => {
+  const response = `worker response to ${data}`;
+  postMessage(response);
+});
